incus-osd/applications: Cleaup and reorganize application code
* Move generic helper methods to their own file
* Sort application methods by alphabetical order
* Move incus-ceph and incus-linstor applications to their own file
