+++ incus-osd/internal/applications/app_migration_manager.go
@@ -27,42 +27,127 @@ type migrationManager struct {
 	common
 }
 
-func (*migrationManager) Name() string {
-	return "migration-manager"
-}
-
-// Start starts the systemd unit.
-func (mm *migrationManager) Start(ctx context.Context) error {
-	err := mm.ConfigureLocalStorage(ctx)
+// AddTrustedCertificate adds a new trusted certificate to the application.
+func (*migrationManager) AddTrustedCertificate(ctx context.Context, _ string, cert string) error {
+	// Compute the certificate's fingerprint.
+	fp, err := getCertificateFingerprint(cert)
 	if err != nil {
 		return err
 	}
 
-	// Start the unit.
-	return systemd.StartUnit(ctx, "migration-manager.service")
-}
-
-// Stop stops the systemd unit.
-func (*migrationManager) Stop(ctx context.Context) error {
-	// Stop the unit.
-	return systemd.StopUnit(ctx, "migration-manager.service")
-}
-
-// Restart restarts the main systemd unit.
-func (*migrationManager) Restart(ctx context.Context) error {
-	return systemd.RestartUnit(ctx, "migration-manager.service")
-}
-
-// Update triggers restart after an application update.
-func (*migrationManager) Update(ctx context.Context) error {
-	// Reload the systemd daemon to pickup any service definition changes.
-	err := systemd.ReloadDaemon(ctx)
+	// Get the current security configuration.
+	body, err := doMMRequest(ctx, "http://localhost/1.0/system/security", http.MethodGet, nil)
 	if err != nil {
 		return err
 	}
 
-	// Restart the unit.
-	return systemd.RestartUnit(ctx, "migration-manager.service")
+	sec := &mmapi.SystemSecurity{}
+
+	err = json.Unmarshal(body, sec)
+	if err != nil {
+		return err
+	}
+
+	// Check if the certificate is already trusted.
+	if slices.Contains(sec.TrustedTLSClientCertFingerprints, fp) {
+		return errors.New("client certificate is already trusted")
+	}
+
+	// Add the certificate's fingerprint to list of trusted clients.
+	sec.TrustedTLSClientCertFingerprints = append(sec.TrustedTLSClientCertFingerprints, fp)
+
+	contentJSON, err := json.Marshal(sec)
+	if err != nil {
+		return err
+	}
+
+	_, err = doMMRequest(ctx, "http://localhost/1.0/system/security", http.MethodPut, contentJSON)
+
+	return err
+}
+
+// ConfigureLocalStorage configures local storage for the application.
+func (mm *migrationManager) ConfigureLocalStorage(ctx context.Context) error {
+	// If the application isn't initialized, create a ZFS dataset for it to use.
+	if !mm.state.Applications["migration-manager"].State.Initialized {
+		err := zfs.CreateApplicationDataset(ctx, "migration-manager")
+		if err != nil {
+			return err
+		}
+	} else {
+		err := zfs.MountApplicationDataset(ctx, "migration-manager")
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// FactoryReset performs a full factory reset of the application.
+func (mm *migrationManager) FactoryReset(ctx context.Context) error {
+	// Stop the application.
+	err := mm.Stop(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Wipe local configuration.
+	err = mm.WipeLocalData(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Start the application.
+	err = mm.Start(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Perform first start initialization.
+	return mm.Initialize(ctx)
+}
+
+// GetBackup returns a tar archive backup of the application's configuration and/or state.
+func (*migrationManager) GetBackup(archive io.Writer, complete bool) error {
+	if complete {
+		return createTarArchive("/var/lib/migration-manager/", nil, archive)
+	}
+
+	return createTarArchive("/var/lib/migration-manager/", []string{"artifacts"}, archive)
+}
+
+// GetClientCertificate returns the keypair for the client certificate.
+func (mm *migrationManager) GetClientCertificate() (*tls.Certificate, error) {
+	// Migration Manager doesn't have a separate certificate it uses when contacting other servers.
+	return mm.GetServerCertificate()
+}
+
+// GetDependencies returns a list of other applications this application depends on.
+func (*migrationManager) GetDependencies() []string {
+	return nil
+}
+
+// GetServerCertificate returns the keypair for the server certificate.
+func (*migrationManager) GetServerCertificate() (*tls.Certificate, error) {
+	// Load the certificate.
+	tlsCert, err := os.ReadFile("/var/lib/migration-manager/server.crt")
+	if err != nil {
+		return nil, err
+	}
+
+	tlsKey, err := os.ReadFile("/var/lib/migration-manager/server.key")
+	if err != nil {
+		return nil, err
+	}
+
+	// Put together a keypair.
+	cert, err := tls.X509KeyPair(tlsCert, tlsKey)
+	if err != nil {
+		return nil, err
+	}
+
+	return &cert, nil
 }
 
 // Initialize runs first time initialization.
@@ -174,9 +259,18 @@ func (mm *migrationManager) Initialize(ctx context.Context) error {
 	return nil
 }
 
+// IsPrimary reports if the application is a primary application.
+func (*migrationManager) IsPrimary() bool {
+	return true
+}
+
 // IsRunning reports if the application is currently running.
 func (*migrationManager) IsRunning(ctx context.Context) bool {
 	return systemd.IsActive(ctx, "migration-manager.service")
+}
+
+func (*migrationManager) Name() string {
+	return "migration-manager"
 }
 
 // NeedsLateUpdateCheck reports if the application depends on a delayed provider update check.
@@ -185,110 +279,43 @@ func (*migrationManager) NeedsLateUpdateCheck() bool {
 	return true
 }
 
-// GetClientCertificate returns the keypair for the client certificate.
-func (mm *migrationManager) GetClientCertificate() (*tls.Certificate, error) {
-	// Migration Manager doesn't have a separate certificate it uses when contacting other servers.
-	return mm.GetServerCertificate()
+// Restart restarts the main systemd unit.
+func (*migrationManager) Restart(ctx context.Context) error {
+	return systemd.RestartUnit(ctx, "migration-manager.service")
 }
 
-// GetServerCertificate returns the keypair for the server certificate.
-func (*migrationManager) GetServerCertificate() (*tls.Certificate, error) {
-	// Load the certificate.
-	tlsCert, err := os.ReadFile("/var/lib/migration-manager/server.crt")
-	if err != nil {
-		return nil, err
-	}
-
-	tlsKey, err := os.ReadFile("/var/lib/migration-manager/server.key")
-	if err != nil {
-		return nil, err
-	}
-
-	// Put together a keypair.
-	cert, err := tls.X509KeyPair(tlsCert, tlsKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return &cert, nil
+// RestoreBackup restores a tar archive backup of the application's configuration and/or state.
+func (*migrationManager) RestoreBackup(ctx context.Context, archive io.Reader) error {
+	return extractTarArchive(ctx, "/var/lib/migration-manager/", []string{"migration-manager.service"}, archive)
 }
 
-// GetDependencies returns a list of other applications this application depends on.
-func (*migrationManager) GetDependencies() []string {
-	return nil
+// Start starts the systemd unit.
+func (mm *migrationManager) Start(ctx context.Context) error {
+	err := mm.ConfigureLocalStorage(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Start the unit.
+	return systemd.StartUnit(ctx, "migration-manager.service")
 }
 
-// AddTrustedCertificate adds a new trusted certificate to the application.
-func (*migrationManager) AddTrustedCertificate(ctx context.Context, _ string, cert string) error {
-	// Compute the certificate's fingerprint.
-	fp, err := getCertificateFingerprint(cert)
-	if err != nil {
-		return err
-	}
-
-	// Get the current security configuration.
-	body, err := doMMRequest(ctx, "http://localhost/1.0/system/security", http.MethodGet, nil)
-	if err != nil {
-		return err
-	}
-
-	sec := &mmapi.SystemSecurity{}
-
-	err = json.Unmarshal(body, sec)
-	if err != nil {
-		return err
-	}
-
-	// Check if the certificate is already trusted.
-	if slices.Contains(sec.TrustedTLSClientCertFingerprints, fp) {
-		return errors.New("client certificate is already trusted")
-	}
-
-	// Add the certificate's fingerprint to list of trusted clients.
-	sec.TrustedTLSClientCertFingerprints = append(sec.TrustedTLSClientCertFingerprints, fp)
-
-	contentJSON, err := json.Marshal(sec)
-	if err != nil {
-		return err
-	}
-
-	_, err = doMMRequest(ctx, "http://localhost/1.0/system/security", http.MethodPut, contentJSON)
-
-	return err
+// Stop stops the systemd unit.
+func (*migrationManager) Stop(ctx context.Context) error {
+	// Stop the unit.
+	return systemd.StopUnit(ctx, "migration-manager.service")
 }
 
-// Migration Manager specific helper to interact with the REST API.
-func doMMRequest(ctx context.Context, url string, method string, body []byte) ([]byte, error) {
-	return doRequest(ctx, "/run/migration-manager/unix.socket", url, method, body)
-}
-
-// IsPrimary reports if the application is a primary application.
-func (*migrationManager) IsPrimary() bool {
-	return true
-}
-
-// FactoryReset performs a full factory reset of the application.
-func (mm *migrationManager) FactoryReset(ctx context.Context) error {
-	// Stop the application.
-	err := mm.Stop(ctx)
+// Update triggers restart after an application update.
+func (*migrationManager) Update(ctx context.Context) error {
+	// Reload the systemd daemon to pickup any service definition changes.
+	err := systemd.ReloadDaemon(ctx)
 	if err != nil {
 		return err
 	}
 
-	// Wipe local configuration.
-	err = mm.WipeLocalData(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Start the application.
-	err = mm.Start(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Perform first start initialization.
-	return mm.Initialize(ctx)
+	// Restart the unit.
+	return systemd.RestartUnit(ctx, "migration-manager.service")
 }
 
 // WipeLocalData removes local data created by the application.
@@ -315,34 +342,7 @@ func (*migrationManager) WipeLocalData(ctx context.Context) error {
 	return zfs.CreateApplicationDataset(ctx, "migration-manager")
 }
 
-// GetBackup returns a tar archive backup of the application's configuration and/or state.
-func (*migrationManager) GetBackup(archive io.Writer, complete bool) error {
-	if complete {
-		return createTarArchive("/var/lib/migration-manager/", nil, archive)
-	}
-
-	return createTarArchive("/var/lib/migration-manager/", []string{"artifacts"}, archive)
-}
-
-// RestoreBackup restores a tar archive backup of the application's configuration and/or state.
-func (*migrationManager) RestoreBackup(ctx context.Context, archive io.Reader) error {
-	return extractTarArchive(ctx, "/var/lib/migration-manager/", []string{"migration-manager.service"}, archive)
-}
-
-// ConfigureLocalStorage configures local storage for the application.
-func (mm *migrationManager) ConfigureLocalStorage(ctx context.Context) error {
-	// If the application isn't initialized, create a ZFS dataset for it to use.
-	if !mm.state.Applications["migration-manager"].State.Initialized {
-		err := zfs.CreateApplicationDataset(ctx, "migration-manager")
-		if err != nil {
-			return err
-		}
-	} else {
-		err := zfs.MountApplicationDataset(ctx, "migration-manager")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
+// Migration Manager specific helper to interact with the REST API.
+func doMMRequest(ctx context.Context, url string, method string, body []byte) ([]byte, error) {
+	return doRequest(ctx, "/run/migration-manager/unix.socket", url, method, body)
 }
