+++ incus-osd/internal/applications/app_operations_center.go
@@ -27,42 +27,133 @@ type operationsCenter struct {
 	common
 }
 
-func (*operationsCenter) Name() string {
-	return "operations-center"
-}
-
-// Start starts the systemd unit.
-func (oc *operationsCenter) Start(ctx context.Context) error {
-	err := oc.ConfigureLocalStorage(ctx)
+// AddTrustedCertificate adds a new trusted certificate to the application.
+func (*operationsCenter) AddTrustedCertificate(ctx context.Context, _ string, cert string) error {
+	// Compute the certificate's fingerprint.
+	fp, err := getCertificateFingerprint(cert)
 	if err != nil {
 		return err
 	}
 
-	// Start the unit.
-	return systemd.StartUnit(ctx, "operations-center.service")
-}
-
-// Stop stops the systemd unit.
-func (*operationsCenter) Stop(ctx context.Context) error {
-	// Stop the unit.
-	return systemd.StopUnit(ctx, "operations-center.service")
-}
-
-// Restart restarts the main systemd unit.
-func (*operationsCenter) Restart(ctx context.Context) error {
-	return systemd.RestartUnit(ctx, "operations-center.service")
-}
-
-// Update triggers restart after an application update.
-func (*operationsCenter) Update(ctx context.Context) error {
-	// Reload the systemd daemon to pickup any service definition changes.
-	err := systemd.ReloadDaemon(ctx)
+	// Get the current security configuration.
+	body, err := doOCRequest(ctx, "http://localhost/1.0/system/security", http.MethodGet, nil)
 	if err != nil {
 		return err
 	}
 
-	// Restart the unit.
-	return systemd.RestartUnit(ctx, "operations-center.service")
+	sec := &ocapi.Security{}
+
+	err = json.Unmarshal(body, sec)
+	if err != nil {
+		return err
+	}
+
+	// Check if the certificate is already trusted.
+	if slices.Contains(sec.TrustedTLSClientCertFingerprints, fp) {
+		return errors.New("client certificate is already trusted")
+	}
+
+	// Add the certificate's fingerprint to list of trusted clients.
+	sec.TrustedTLSClientCertFingerprints = append(sec.TrustedTLSClientCertFingerprints, fp)
+
+	contentJSON, err := json.Marshal(sec)
+	if err != nil {
+		return err
+	}
+
+	_, err = doOCRequest(ctx, "http://localhost/1.0/system/security", http.MethodPut, contentJSON)
+
+	return err
+}
+
+// ConfigureLocalStorage configures local storage for the application.
+func (oc *operationsCenter) ConfigureLocalStorage(ctx context.Context) error {
+	// If the application isn't initialized, create a ZFS dataset for it to use.
+	if !oc.state.Applications["operations-center"].State.Initialized {
+		err := zfs.CreateApplicationDataset(ctx, "operations-center")
+		if err != nil {
+			return err
+		}
+	} else {
+		err := zfs.MountApplicationDataset(ctx, "operations-center")
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// FactoryReset performs a full factory reset of the application.
+func (oc *operationsCenter) FactoryReset(ctx context.Context) error {
+	// Stop the application.
+	err := oc.Stop(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Wipe local configuration.
+	err = oc.WipeLocalData(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Check if we're locally registered with it.
+	if oc.state.System.Provider.State.Registered && (oc.state.System.Provider.Config.Config == nil || oc.state.System.Provider.Config.Config["server_url"] == "") {
+		oc.state.System.Provider.State.Registered = false
+		oc.state.System.Provider.Config.Config = map[string]string{}
+	}
+
+	// Start the application.
+	err = oc.Start(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Perform first start initialization.
+	return oc.Initialize(ctx)
+}
+
+// GetBackup returns a tar archive backup of the application's configuration and/or state.
+func (*operationsCenter) GetBackup(archive io.Writer, complete bool) error {
+	if complete {
+		return createTarArchive("/var/lib/operations-center/", nil, archive)
+	}
+
+	return createTarArchive("/var/lib/operations-center/", []string{"updates"}, archive)
+}
+
+// GetClientCertificate returns the keypair for the client certificate.
+func (oc *operationsCenter) GetClientCertificate() (*tls.Certificate, error) {
+	// Operations Center doesn't have a separate certificate it uses when contacting other servers.
+	return oc.GetServerCertificate()
+}
+
+// GetDependencies returns a list of other applications this application depends on.
+func (*operationsCenter) GetDependencies() []string {
+	return nil
+}
+
+// GetServerCertificate returns the keypair for the server certificate.
+func (*operationsCenter) GetServerCertificate() (*tls.Certificate, error) {
+	// Load the certificate.
+	tlsCert, err := os.ReadFile("/var/lib/operations-center/server.crt")
+	if err != nil {
+		return nil, err
+	}
+
+	tlsKey, err := os.ReadFile("/var/lib/operations-center/server.key")
+	if err != nil {
+		return nil, err
+	}
+
+	// Put together a keypair.
+	cert, err := tls.X509KeyPair(tlsCert, tlsKey)
+	if err != nil {
+		return nil, err
+	}
+
+	return &cert, nil
 }
 
 // Initialize runs first time initialization.
@@ -189,9 +280,18 @@ func (oc *operationsCenter) Initialize(ctx context.Context) error {
 	return nil
 }
 
+// IsPrimary reports if the application is a primary application.
+func (*operationsCenter) IsPrimary() bool {
+	return true
+}
+
 // IsRunning reports if the application is currently running.
 func (*operationsCenter) IsRunning(ctx context.Context) bool {
 	return systemd.IsActive(ctx, "operations-center.service")
+}
+
+func (*operationsCenter) Name() string {
+	return "operations-center"
 }
 
 // NeedsLateUpdateCheck reports if the application depends on a delayed provider update check.
@@ -202,116 +302,43 @@ func (*operationsCenter) NeedsLateUpdateCheck() bool {
 	return true
 }
 
-// GetClientCertificate returns the keypair for the client certificate.
-func (oc *operationsCenter) GetClientCertificate() (*tls.Certificate, error) {
-	// Operations Center doesn't have a separate certificate it uses when contacting other servers.
-	return oc.GetServerCertificate()
+// Restart restarts the main systemd unit.
+func (*operationsCenter) Restart(ctx context.Context) error {
+	return systemd.RestartUnit(ctx, "operations-center.service")
 }
 
-// GetServerCertificate returns the keypair for the server certificate.
-func (*operationsCenter) GetServerCertificate() (*tls.Certificate, error) {
-	// Load the certificate.
-	tlsCert, err := os.ReadFile("/var/lib/operations-center/server.crt")
-	if err != nil {
-		return nil, err
-	}
-
-	tlsKey, err := os.ReadFile("/var/lib/operations-center/server.key")
-	if err != nil {
-		return nil, err
-	}
-
-	// Put together a keypair.
-	cert, err := tls.X509KeyPair(tlsCert, tlsKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return &cert, nil
+// RestoreBackup restores a tar archive backup of the application's configuration and/or state.
+func (*operationsCenter) RestoreBackup(ctx context.Context, archive io.Reader) error {
+	return extractTarArchive(ctx, "/var/lib/operations-center/", []string{"operations-center.service"}, archive)
 }
 
-// GetDependencies returns a list of other applications this application depends on.
-func (*operationsCenter) GetDependencies() []string {
-	return nil
+// Start starts the systemd unit.
+func (oc *operationsCenter) Start(ctx context.Context) error {
+	err := oc.ConfigureLocalStorage(ctx)
+	if err != nil {
+		return err
+	}
+
+	// Start the unit.
+	return systemd.StartUnit(ctx, "operations-center.service")
 }
 
-// AddTrustedCertificate adds a new trusted certificate to the application.
-func (*operationsCenter) AddTrustedCertificate(ctx context.Context, _ string, cert string) error {
-	// Compute the certificate's fingerprint.
-	fp, err := getCertificateFingerprint(cert)
-	if err != nil {
-		return err
-	}
-
-	// Get the current security configuration.
-	body, err := doOCRequest(ctx, "http://localhost/1.0/system/security", http.MethodGet, nil)
-	if err != nil {
-		return err
-	}
-
-	sec := &ocapi.Security{}
-
-	err = json.Unmarshal(body, sec)
-	if err != nil {
-		return err
-	}
-
-	// Check if the certificate is already trusted.
-	if slices.Contains(sec.TrustedTLSClientCertFingerprints, fp) {
-		return errors.New("client certificate is already trusted")
-	}
-
-	// Add the certificate's fingerprint to list of trusted clients.
-	sec.TrustedTLSClientCertFingerprints = append(sec.TrustedTLSClientCertFingerprints, fp)
-
-	contentJSON, err := json.Marshal(sec)
-	if err != nil {
-		return err
-	}
-
-	_, err = doOCRequest(ctx, "http://localhost/1.0/system/security", http.MethodPut, contentJSON)
-
-	return err
+// Stop stops the systemd unit.
+func (*operationsCenter) Stop(ctx context.Context) error {
+	// Stop the unit.
+	return systemd.StopUnit(ctx, "operations-center.service")
 }
 
-// Operations Center specific helper to interact with the REST API.
-func doOCRequest(ctx context.Context, url string, method string, body []byte) ([]byte, error) {
-	return doRequest(ctx, "/run/operations-center/unix.socket", url, method, body)
-}
-
-// IsPrimary reports if the application is a primary application.
-func (*operationsCenter) IsPrimary() bool {
-	return true
-}
-
-// FactoryReset performs a full factory reset of the application.
-func (oc *operationsCenter) FactoryReset(ctx context.Context) error {
-	// Stop the application.
-	err := oc.Stop(ctx)
+// Update triggers restart after an application update.
+func (*operationsCenter) Update(ctx context.Context) error {
+	// Reload the systemd daemon to pickup any service definition changes.
+	err := systemd.ReloadDaemon(ctx)
 	if err != nil {
 		return err
 	}
 
-	// Wipe local configuration.
-	err = oc.WipeLocalData(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Check if we're locally registered with it.
-	if oc.state.System.Provider.State.Registered && (oc.state.System.Provider.Config.Config == nil || oc.state.System.Provider.Config.Config["server_url"] == "") {
-		oc.state.System.Provider.State.Registered = false
-		oc.state.System.Provider.Config.Config = map[string]string{}
-	}
-
-	// Start the application.
-	err = oc.Start(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Perform first start initialization.
-	return oc.Initialize(ctx)
+	// Restart the unit.
+	return systemd.RestartUnit(ctx, "operations-center.service")
 }
 
 // WipeLocalData removes local data created by the application.
@@ -338,34 +365,7 @@ func (*operationsCenter) WipeLocalData(ctx context.Context) error {
 	return zfs.CreateApplicationDataset(ctx, "operations-center")
 }
 
-// GetBackup returns a tar archive backup of the application's configuration and/or state.
-func (*operationsCenter) GetBackup(archive io.Writer, complete bool) error {
-	if complete {
-		return createTarArchive("/var/lib/operations-center/", nil, archive)
-	}
-
-	return createTarArchive("/var/lib/operations-center/", []string{"updates"}, archive)
-}
-
-// RestoreBackup restores a tar archive backup of the application's configuration and/or state.
-func (*operationsCenter) RestoreBackup(ctx context.Context, archive io.Reader) error {
-	return extractTarArchive(ctx, "/var/lib/operations-center/", []string{"operations-center.service"}, archive)
-}
-
-// ConfigureLocalStorage configures local storage for the application.
-func (oc *operationsCenter) ConfigureLocalStorage(ctx context.Context) error {
-	// If the application isn't initialized, create a ZFS dataset for it to use.
-	if !oc.state.Applications["operations-center"].State.Initialized {
-		err := zfs.CreateApplicationDataset(ctx, "operations-center")
-		if err != nil {
-			return err
-		}
-	} else {
-		err := zfs.MountApplicationDataset(ctx, "operations-center")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
+// Operations Center specific helper to interact with the REST API.
+func doOCRequest(ctx context.Context, url string, method string, body []byte) ([]byte, error) {
+	return doRequest(ctx, "/run/operations-center/unix.socket", url, method, body)
 }
